requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM and AI
openai>=1.12.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5

# Async support
aioredis>=2.0.1
asyncio-redis>=0.16.0

# Monitoring and logging
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-redis>=0.43b0

# Testing and API client
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-redis>=3.0.2
httpx>=0.26.0

# Development
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
